apply previous fixes, which removed all warnings and appears to have fixed video export on windows

+++ Basics/foleys_VideoFifo.cpp
@@ -39,11 +39,11 @@ std::pair<int64_t, juce::Image> VideoFifo::popVideoFrame()
 
 std::pair<int64_t, juce::Image> VideoFifo::getVideoFrame (double timestamp) const
 {
-    const juce::int64 pts = timestamp * settings.timebase;
+    const juce::int64 pts = (int64_t)(timestamp * settings.timebase);
 
     const juce::ScopedLock sl (lock);
 
-    auto vf = videoFrames.lower_bound (timestamp * settings.timebase);
+    auto vf = videoFrames.lower_bound ((int64_t)(timestamp * settings.timebase));
     if (vf == videoFrames.end())
         return {};
 
@@ -56,7 +56,7 @@ std::pair<int64_t, juce::Image> VideoFifo::getVideoFrame (double timestamp) cons
 
 bool VideoFifo::isFrameAvailable (double timestamp) const
 {
-    const juce::int64 pts = timestamp * settings.timebase;
+    const juce::int64 pts = (int64_t)(timestamp * settings.timebase);
 
     const juce::ScopedLock sl (lock);
 
@@ -72,7 +72,7 @@ bool VideoFifo::isFrameAvailable (double timestamp) const
 
 int64_t VideoFifo::getFrameCountForTime (double timestamp) const
 {
-    const juce::int64 pts = timestamp * settings.timebase;
+    const juce::int64 pts = (int64_t)(timestamp * settings.timebase);
 
     const juce::ScopedLock sl (lock);
